Move SimpleQualityMetric into a subdirectory
